assemble tui
